% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HttpSampler.R
\name{rHttpSampler}
\alias{rHttpSampler}
\title{Sample Random HTTP Requests from a REST API}
\usage{
rHttpSampler(n = 5, baseUrl, swaggerUrl)
}
\arguments{
\item{n}{Integer. Number of sampling iterations (default: 5).}

\item{baseUrl}{Character. Base URL of the target REST API service.}

\item{swaggerUrl}{Character. URL of the OpenAPI/Swagger JSON specification.}
}
\value{
A data frame containing information about sampled requests
and responses.
}
\description{
The `rHttpSampler()` function provides a lightweight interface for
black-box REST API testing using a Kotlin-based sampling engine derived
from the EvoMaster framework. It executes random HTTP requests defined
in an OpenAPI specification and returns structured response data as a
data frame.
}
\examples{
\dontrun{
results_df <- rHttpSampler(
  n = 5,
  baseUrl = "https://petstore.swagger.io",
  swaggerUrl = "https://petstore.swagger.io/v2/swagger.json"
)
View(results_df)
}

}
\references{
Arcuri, Andrea. "RESTful API automated test case generation with EvoMaster." ACM Transactions on Software Engineering and Methodology (TOSEM) 28.1 (2019): 1-37.
\doi{doi:10.1145/3293455}

Arcuri, Andrea, et al. "Evomaster: A search-based system test generation tool." (2021).
\doi{doi:10.21105/joss.02153}
}
